extracted asset airdrop into a seperate component
