+++ src/components/airdrops/AssetAirdrop/index.ts
@@ -1,0 +1,3 @@
+import AssetAirdrop from './AssetAirdrop.vue';
+
+export default AssetAirdrop;
