+++ src/views/Asset.vue
@@ -91,44 +91,7 @@
         </section>
 
         <!-- Token Airdrop -->
-        <div v-if="showAirdropCard" class="bg-fg w-full p-4 flex items-center justify-between rounded-xl mt-8">
-          <div class="flex items-center">
-            <img src="~@/assets/images/token-airdrop.png" alt="Token Airdrop" class="w-8 h-8 rounded-full mr-2" />
-            <p>
-              <span class="uppercase">{{ assetAirdrop[0].project }}</span> {{ $t('context.airdrops.airdrop') }}
-            </p>
-          </div>
-          <div class="text-muted">0.123 {{ assetAirdrop[0].tokenTicker }}</div>
-          <div class="flex items-center">
-            <div class="mr-6">
-              <Button
-                v-if="assetAirdrop[0].eligibility === AirdropEligibilityStatus.CLAIMABLE"
-                name="Claim"
-                size="sm"
-                variant="secondary"
-              />
-              <div v-else-if="assetAirdrop[0].eligibility === AirdropEligibilityStatus.ELIGIBLE">Eligible</div>
-              <div
-                v-if="assetAirdrop[0].eligibility === AirdropEligibilityStatus.AUTO_DROP"
-                class="flex items-center float-right"
-              >
-                <Icon :name="'CheckIcon'" :icon-size="1" class="mr-2" />Auto-drop
-              </div>
-              <div
-                v-else-if="assetAirdrop[0].eligibility === AirdropEligibilityStatus.CLAIMED"
-                class="flex items-center float-right"
-              >
-                <Icon :name="'CheckIcon'" :icon-size="1" class="mr-2" />Claimed
-              </div>
-            </div>
-            <Icon
-              :name="'CloseIcon'"
-              :icon-size="1"
-              class="cursor-pointer"
-              @click="() => (airdropCardIsVisible = false)"
-            />
-          </div>
-        </div>
+        <AssetAirdrop :denom="denom" />
 
         <!-- Staking -->
         <template v-if="stakingEnabled">
@@ -220,11 +183,12 @@
 </template>
 
 <script lang="ts">
-import { computed, defineComponent, onMounted, onUnmounted, ref, watch } from 'vue';
+import { computed, defineComponent, onUnmounted, ref, watch } from 'vue';
 import { useMeta } from 'vue-meta';
 import { useRoute, useRouter } from 'vue-router';
 import { useStore } from 'vuex';
 
+import AssetAirdrop from '@/components/airdrops/AssetAirdrop';
 import StakeTable from '@/components/asset/StakeTable.vue';
 import PoolBanner from '@/components/assets/AssetsTable/PoolBanner.vue';
 // import StakingBanner from '@/components/banners/StakingBanner.vue';
@@ -242,15 +206,12 @@ import Pools from '@/components/liquidity/Pools.vue';
 import LiquiditySwap from '@/components/liquidity/Swap.vue';
 import TooltipPools from '@/components/liquidity/TooltipPools.vue';
 import DexSwap from '@/components/swap/DexSwap.vue';
-import Button from '@/components/ui/Button.vue';
-import Icon from '@/components/ui/Icon.vue';
 import useAccount from '@/composables/useAccount';
 import usePools from '@/composables/usePools';
 import AppLayout from '@/layouts/AppLayout.vue';
 import { GlobalActionTypes, GlobalGetterTypes, RootStoreTyped } from '@/store';
 import { LoadingState } from '@/types/util';
 import { getDisplayName, getTicker } from '@/utils/actionHandler';
-import { AirdropEligibilityStatus } from '@/utils/airdropEligibility';
 import { pageview } from '@/utils/analytics';
 import { generateDenomHash, parseCoins } from '@/utils/basic';
 import { featureRunning } from '@/utils/FeatureManager';
@@ -276,16 +237,13 @@ export default defineComponent({
     ChainDownWarning,
     AreaChart,
     DexSwap,
-    Icon,
-    Button,
     FeatureRunningConditional,
+    AssetAirdrop,
   },
 
   setup() {
     const displayName = ref('');
-    const tokenTicker = ref('');
     const displayPrice = ref(0);
-    const airdropCardIsVisible = ref(true);
     const metaSource = computed(() => {
       return { title: displayName.value };
     });
@@ -481,10 +439,11 @@ export default defineComponent({
       });
 
       getTokenPrices.value = async (days: string, showSkeleton: boolean) => {
+        const tokenTicker = await getTicker(denom.value, typedstore.getters[GlobalGetterTypes.API.getDexChain]);
         const chainName = await typedstore.dispatch(GlobalActionTypes.API.GET_COINGECKO_ID_BY_NAMES, {
           subscribe: false,
           params: {
-            token: tokenTicker.value.toLowerCase(),
+            token: tokenTicker.toLowerCase(),
             showSkeleton: false,
           },
         });
@@ -521,37 +480,7 @@ export default defineComponent({
       typedstore.dispatch(GlobalActionTypes.API.RESET_TOKEN_PRICES);
     });
 
-    onMounted(async () => {
-      tokenTicker.value = await getTicker(denom.value, typedstore.getters[GlobalGetterTypes.API.getDexChain]);
-    });
-
     const isStakingRunning = featureRunning('STAKING');
-    const isAirdropsRunning = featureRunning('AIRDROPS_FEATURE');
-
-    const assetAirdrop = computed(() => {
-      const airdrops = typedstore.getters[GlobalGetterTypes.API.getAirdrops];
-      return airdrops.filter((item) => item.tokenTicker === tokenTicker.value);
-    });
-
-    const isDemoAccount = computed(() => {
-      return (
-        !typedstore.getters[GlobalGetterTypes.USER.isSignedIn] ||
-        typedstore.getters[GlobalGetterTypes.USER.isDemoAccount]
-      );
-    });
-
-    const showAirdropCard = computed(() => {
-      console.log('checking here', assetAirdrop.value);
-      return (
-        isAirdropsRunning &&
-        !isDemoAccount.value &&
-        airdropCardIsVisible.value &&
-        assetAirdrop.value[0] &&
-        assetAirdrop.value[0].eligibility !== AirdropEligibilityStatus.ENDED &&
-        assetAirdrop.value[0].eligibility !== AirdropEligibilityStatus.NOT_ELIGIBLE &&
-        assetAirdrop.value[0].eligibility !== AirdropEligibilityStatus.NOT_AVAILABLE
-      );
-    });
 
     return {
       nativeAsset,
@@ -575,15 +504,8 @@ export default defineComponent({
       priceDiffObject,
       setPriceDifference,
       isStakingRunning,
-      isAirdropsRunning,
       displayPrice,
-      assetAirdrop,
-      showAirdropCard,
-      airdropCardIsVisible,
-      AirdropEligibilityStatus,
     };
   },
 });
 </script>
-
-<style lang="scss" scoped></style>
